Finished testing of service.
Added testing for the rest of the service, including the Service, the storage, the Credentials and other Model classes.
Additionally, added hashcode to the model classes for better functionality with HashMap storage.
Added in ErrorMessages class for use wth all error messages within backend.
Finally, added Credentials and CredentialsStorage for use in Gateway API for connection to FrontEnd.