Added filter ui
Added the UI for the implementation of filters, also with the implementation on the backend. Has a wierd message in console: class java.lang.String cannot be cast to class devops.model.implementations.PersonNetwork (java.lang.String is in module java.base of loader 'bootstrap'; devops.model.implementations.PersonNetwork is in unnamed module of loader 'app')

Finally, removed handlers for textboxes in information to side panel.